requirements.txt
crowdstrike-falconpy==0.4.8
deb-pkg-tools==8.4
.
